Revert "created filteration based on topic for mypage (#83)"
This reverts commit 1e8497ce110e7ec93a9448b0d67ca651265104af.
